Update to angular-ui-bootstrap
Was previously referencing bootstrap.js, which in turn required jQuery.
angular-ui-bootstrap is a project that removes the requirement for these
dependencies, with smaller file size to boot!

I'm not sure it's a great idea to reference the script off github direct rather
than a CDN but I couldn't find a CDN for the project. An alternative is
downloading the code and serving it direct from the application server but this
seems wasteful.